requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-cognitiveservices-speech==1.38.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Login==0.6.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-language==2.13.3
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.68.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openai==1.20.0
packaging==24.2
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
